+++ xds/src/test/java/io/grpc/xds/XdsLoadStatsStoreTest.java
@@ -17,6 +17,8 @@
 package io.grpc.xds;
 
 import static com.google.common.truth.Truth.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -27,6 +29,11 @@ import io.envoyproxy.envoy.api.v2.endpoint.ClusterStats;
 import io.envoyproxy.envoy.api.v2.endpoint.ClusterStats.DroppedRequests;
 import io.envoyproxy.envoy.api.v2.endpoint.EndpointLoadMetricStats;
 import io.envoyproxy.envoy.api.v2.endpoint.UpstreamLocalityStats;
+import io.grpc.ClientStreamTracer;
+import io.grpc.LoadBalancer.PickResult;
+import io.grpc.LoadBalancer.Subchannel;
+import io.grpc.Metadata;
+import io.grpc.Status;
 import io.grpc.xds.ClientLoadCounter.ClientLoadSnapshot;
 import io.grpc.xds.ClientLoadCounter.MetricValue;
 import io.grpc.xds.XdsLoadStatsStore.StatsCounter;
@@ -63,6 +70,9 @@ public class XdsLoadStatsStoreTest {
           .setZone("test_zone")
           .setSubZone("test_subzone")
           .build();
+  private static final ClientStreamTracer.StreamInfo STREAM_INFO =
+      ClientStreamTracer.StreamInfo.newBuilder().build();
+  private Subchannel mockSubchannel = mock(Subchannel.class);
   private ConcurrentMap<Locality, StatsCounter> localityLoadCounters;
   private ConcurrentMap<String, AtomicLong> dropCounters;
   private XdsLoadStatsStore loadStore;
@@ -295,5 +305,42 @@ public class XdsLoadStatsStoreTest {
     assertClusterStatsEqual(expectedLoadReport, loadStore.generateLoadReport());
     assertThat(dropCounters.get("lb").get()).isEqualTo(0);
     assertThat(dropCounters.get("throttle").get()).isEqualTo(0);
+  }
+
+  @Test
+  public void loadNotRecordedForUntrackedLocality() {
+    PickResult pickResult = PickResult.withSubchannel(mockSubchannel);
+    // If the per-locality counter does not exist, nothing should happen.
+    PickResult interceptedPickResult = loadStore.interceptPickResult(pickResult, LOCALITY1);
+    assertThat(interceptedPickResult.getStreamTracerFactory()).isNull();
+  }
+
+  @Test
+  public void invalidPickResultNotIntercepted() {
+    PickResult errorResult = PickResult.withError(Status.UNAVAILABLE.withDescription("Error"));
+    PickResult droppedResult = PickResult.withDrop(Status.UNAVAILABLE.withDescription("Dropped"));
+    // TODO (chengyuanzhang): for NoResult PickResult, do we still intercept?
+    PickResult interceptedErrorResult = loadStore.interceptPickResult(errorResult, LOCALITY1);
+    PickResult interceptedDroppedResult =
+        loadStore.interceptPickResult(droppedResult, LOCALITY1);
+    assertThat(interceptedErrorResult.getStreamTracerFactory()).isNull();
+    assertThat(interceptedDroppedResult.getStreamTracerFactory()).isNull();
+  }
+
+  @Test
+  public void interceptPreservesOriginStreamTracer() {
+    ClientStreamTracer.Factory mockFactory = mock(ClientStreamTracer.Factory.class);
+    ClientStreamTracer mockTracer = mock(ClientStreamTracer.class);
+    when(mockFactory
+        .newClientStreamTracer(any(ClientStreamTracer.StreamInfo.class), any(Metadata.class)))
+        .thenReturn(mockTracer);
+    localityLoadCounters.put(LOCALITY1, new ClientLoadCounter());
+    PickResult pickResult = PickResult.withSubchannel(mockSubchannel, mockFactory);
+    PickResult interceptedPickResult = loadStore.interceptPickResult(pickResult, LOCALITY1);
+    Metadata metadata = new Metadata();
+    interceptedPickResult.getStreamTracerFactory().newClientStreamTracer(STREAM_INFO, metadata)
+        .streamClosed(Status.OK);
+    verify(mockFactory).newClientStreamTracer(same(STREAM_INFO), same(metadata));
+    verify(mockTracer).streamClosed(Status.OK);
   }
 }
